New file structure in the works
New structure, see previous comments for description, is underway.
Further functionality to come...
