+++ detect_avoid_files/CMakeLists.txt
@@ -13,7 +13,8 @@ catkin_package()
 
 include_directories(include ${catkin_INCLUDE_DIRS})
 
-add_executable(detect_avoid src/test_flight.cpp)
+add_executable(detect_avoid src/avoid.cpp)
+add_executable(detect_avoid src/detect.cpp)
 target_link_libraries(detect_avoid ${catkin_LIBRARIES})
 add_dependencies(dectect_avoid detect_avoid_generate_messages_cpp)
